# internal/web/assets/openapi.yaml
openapi: 3.0.3
info:
  title: PasarGuard Panel API
  version: "1.0.0"
  description: >
    API پنل اصلی برای مدیریت چند نود، مشتری، پلن، اشتراک و سهمیه.
    این API سطح یکپارچه‌سازی برای ربات فروش است. پنل منطق پول/کیف‌پول ندارد؛
    فقط بایت/زمان/وضعیت را مدیریت و مصرف/overage را expose می‌کند.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /sub/{token}:
    get:
      tags: [system]
      summary: Public subscription page (HTML) for the customer — config + quota/expiry
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: HTML page }
        "404": { description: Not found }
  /sub/{token}/info:
    get:
      tags: [system]
      summary: Public subscription info (JSON) — config + quota/expiry
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Info
          content: { application/json: { schema: { type: object } } }
        "404": { description: Not found }
  /api/v1/customers:
    get:
      tags: [customers]
      summary: List customers
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
    post:
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                external_ref: { type: string, description: "شناسه‌ی مشتری در ربات فروش" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
  /api/v1/customers/{id}/usage:
    get:
      tags: [usage]
      summary: Aggregated usage for a customer
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Usage
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerUsage" }
  /api/v1/customers/{id}/subscriptions:
    get:
      tags: [subscriptions]
      summary: List a customer's subscriptions
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Subscription" }
    post:
      tags: [subscriptions]
      summary: Provision a subscription (creates a tenant on a node)
      description: >
        یک tenant روی نود انتخابی می‌سازد و کلید API مشتری را **فقط یک‌بار** برمی‌گرداند.
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id, node_id]
              properties:
                plan_id: { type: string }
                node_id: { type: string }
                credit_limit_bytes: { type: integer, format: int64 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubscriptionResponse" }
        "502": { $ref: "#/components/responses/NodeError" }
  /api/v1/plans:
    get:
      tags: [plans]
      summary: List plans
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Plan" }
    post:
      tags: [plans]
      summary: Create a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, quota_bytes, duration_days]
              properties:
                name: { type: string }
                quota_bytes: { type: integer, format: int64 }
                duration_days: { type: integer }
                max_users: { type: integer }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plan" }
  /api/v1/nodes:
    get:
      tags: [nodes]
      summary: List nodes
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Node" }
    post:
      tags: [nodes]
      summary: Register a node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address, master_key]
              properties:
                name: { type: string }
                address: { type: string, example: "https://1.2.3.4:8090" }
                master_key: { type: string }
                core_key: { type: string, description: "کلید مدیریت هسته از پنل PasarGuard (اختیاری)" }
                grpc_port: { type: integer, description: "پورت gRPC سازگار PasarGuard (پیش‌فرض 62050)" }
                cert_pem: { type: string, description: "گواهی سلف‌ساین نود برای pin (خالی = دریافت خودکار TOFU)" }
                host_info: { type: string, description: "یادداشت آزاد اپراتور درباره‌ی هاست/کانفیگ؛ در پاسخ connection به خریدار داده می‌شود" }
                config: { type: object, description: "کانفیگ ثابت Xray (اختیاری) برای push" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
  /api/v1/nodes/{id}:
    get:
      tags: [nodes]
      summary: Node connection details (host, ports, keys, certificate) for the operator
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Node detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeDetail" }
    patch:
      tags: [nodes]
      summary: Update a node (partial; omitted fields keep their stored value)
      description: >
        هر فیلد اختیاری است. فیلدی که ارسال نشود بدون تغییر می‌ماند؛ ارسال رشته‌ی
        خالی یعنی «پاک کن». name و address را می‌توان حذف کرد ولی نمی‌توان خالی فرستاد.
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                address: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                grpc_port: { type: integer, description: "ارسال‌نشده = بدون تغییر" }
                master_key: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                core_key: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                cert_pem: { type: string, description: "ارسال‌نشده = بدون تغییر" }
                host_info: { type: string, description: "ارسال‌نشده = بدون تغییر" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/Node" } } } }
    delete:
      tags: [nodes]
      summary: Delete a node
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "204": { description: Deleted }
...
  /api/v1/nodes/{id}/core/{action}:
    post:
      tags: [nodes]
      summary: Core lifecycle — action is start|stop|restart
      parameters:
        - { $ref: "#/components/parameters/IdParam" }
        - { name: action, in: path, required: true, schema: { type: string, enum: [start, stop, restart] } }
      responses:
        "200": { description: Done }
        "502": { $ref: "#/components/responses/NodeError" }
...
  /api/v1/nodes/{id}/update:
    post:
      tags: [nodes]
      summary: Self-update the node binary (latest or given version) and restart the service
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { version: { type: string, example: "latest" } } } } }
      responses:
        "200": { description: Updating }
        "502": { $ref: "#/components/responses/NodeError" }
  /api/v1/subscriptions/{id}/suspend:
    post:
      tags: [subscriptions]
      summary: Suspend a subscription (forwarded to the node)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
  /api/v1/subscriptions/{id}/resume:
    post:
      tags: [subscriptions]
      summary: Resume a subscription
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
  /api/v1/subscriptions/{id}/topup-quota:
    post:
      tags: [subscriptions]
      summary: Add quota bytes (and reactivate)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [add_bytes]
              properties:
                add_bytes: { type: integer, format: int64 }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
  /api/v1/webhooks:
    get:
      tags: [webhooks]
      summary: List webhook endpoints
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Webhook" }
    post:
      tags: [webhooks]
      summary: Register a webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, secret]
              properties:
                url: { type: string }
                secret: { type: string, description: "برای امضای HMAC-SHA256" }
                events: { type: string, description: "کاما-جدا یا *", example: "*" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Webhook" }
  /api/v1/customers/{id}:
    delete:
      tags: [customers]
      summary: Delete a customer (deprovisions tenants, cascades subscriptions)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "204": { description: Deleted }
  /api/v1/customers/{id}/enable:
    post:
      tags: [customers]
      summary: Enable a customer (resumes its subscriptions)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Updated
          content: { application/json: { schema: { $ref: "#/components/schemas/Customer" } } }
  /api/v1/customers/{id}/disable:
    post:
      tags: [customers]
      summary: Disable a customer (suspends its subscriptions)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Updated
          content: { application/json: { schema: { $ref: "#/components/schemas/Customer" } } }
  /api/v1/plans/{id}:
    delete:
      tags: [plans]
      summary: Delete a plan
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "204": { description: Deleted }
  /api/v1/nodes/{id}/config:
    get:
      tags: [nodes]
      summary: Get the node's current running Xray config (live, falls back to last pushed)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Config JSON (object)
          content: { application/json: { schema: { type: object } } }
    put:
      tags: [nodes]
      summary: Push a fixed Xray config to the node (applies and restarts core)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "200": { description: Applied }
        "502": { $ref: "#/components/responses/NodeError" }
  /api/v1/nodes/{id}/inbounds:
    get:
      tags: [nodes]
      summary: Customer-shareable inbound definitions of the node
      description: >
        فقط بخش inbounds کانفیگ نود را برمی‌گرداند (بدون outbound/routing) تا
        به مشتری بدهی و دقیقاً همان inbound را در پنل خودش بسازد. اگر force-inbounds
        روی نود تنظیم شده باشد، فقط همان تگ‌ها برگردانده می‌شوند.
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: "Inbounds document: { inbounds: [...] }"
          content: { application/json: { schema: { type: object } } }
        "502": { $ref: "#/components/responses/NodeError" }
  /api/v1/subscriptions/{id}/connection:
    get:
      tags: [subscriptions]
      summary: Connection info to hand to the customer (gRPC address, cert, inbounds)
      description: >
        هرچه مشتری برای افزودن این نود در پنل PasarGuard خودش لازم دارد را برمی‌گرداند:
        آدرس gRPC، گواهی نود و inboundهای قابل‌اشتراک. کلید مشتری اینجا نیست (فقط
        یک‌بار هنگام ساخت اشتراک نمایش داده می‌شود).
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Connection info
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConnectionInfo" }
  /api/v1/subscriptions/{id}:
    delete:
      tags: [subscriptions]
      summary: Delete a subscription (deprovisions the tenant from its node)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "204": { description: Deleted }
  /api/v1/subscriptions/{id}/renew:
    post:
      tags: [subscriptions]
      summary: Renew a subscription (reset usage, extend expiry, reactivate)
      parameters: [{ $ref: "#/components/parameters/IdParam" }]
      responses:
        "200":
          description: Updated
          content: { application/json: { schema: { $ref: "#/components/schemas/Subscription" } } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    NodeError:
      description: خطا در ارتباط با نود
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
    Customer:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        status: { type: string }
        external_ref: { type: string }
        created_at: { type: integer, format: int64 }
    Plan:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        quota_bytes: { type: integer, format: int64 }
        duration_days: { type: integer }
        max_users: { type: integer }
        created_at: { type: integer, format: int64 }
    Node:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        address: { type: string }
        grpc_port: { type: integer, example: 62050 }
        status: { type: string, enum: [online, offline, unknown] }
        version: { type: string }
        capacity_score: { type: integer }
        last_seen_at: { type: integer, format: int64 }
        created_at: { type: integer, format: int64 }
    NodeDetail:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        host: { type: string, example: "1.2.3.4" }
        address: { type: string, example: "https://1.2.3.4:8090" }
        service_port: { type: integer, example: 8090 }
        grpc_address: { type: string, example: "1.2.3.4:62050" }
        grpc_port: { type: integer, example: 62050 }
        protocol: { type: string, example: "grpc" }
        master_key: { type: string }
        core_key: { type: string }
        cert_pem: { type: string }
        cert_pinned:
          type: boolean
          description: >
            false یعنی گواهی معتبری ذخیره نشده و ارتباط پنل با نود بدون تأیید هویت
            انجام می‌شود (در حالی که master key در هر درخواست ارسال می‌گردد).
        host_info:
          type: string
          description: "یادداشت آزاد اپراتور؛ در پاسخ connection به خریدار داده می‌شود"
        status: { type: string }
        version: { type: string }
        last_seen_at: { type: integer, format: int64 }
        created_at: { type: integer, format: int64 }
...
    Subscription:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        plan_id: { type: string }
        node_id: { type: string }
        node_tenant_id: { type: string }
        status: { type: string, enum: [active, suspended, expired] }
        period_id: { type: integer }
        start_at: { type: integer, format: int64 }
        end_at: { type: integer, format: int64 }
        quota_bytes: { type: integer, format: int64 }
        used_bytes: { type: integer, format: int64 }
        credit_limit_bytes: { type: integer, format: int64 }
        notified_threshold: { type: integer }
        created_at: { type: integer, format: int64 }
    SubscriptionResponse:
      type: object
      properties:
        subscription: { $ref: "#/components/schemas/Subscription" }
        api_key: { type: string, description: "کلید مشتری — فقط همین یک‌بار نمایش داده می‌شود" }
        node_address: { type: string }
        sub_token: { type: string, description: "توکن صفحه‌ی عمومی اشتراک (/sub/{token})" }
    CustomerUsage:
      type: object
      properties:
        customer_id: { type: string }
        quota_bytes: { type: integer, format: int64 }
        used_bytes: { type: integer, format: int64 }
        overage_bytes: { type: integer, format: int64 }
    ConnectionInfo:
      type: object
      properties:
        node_name: { type: string }
        grpc_address: { type: string, example: "1.2.3.4:62050" }
        protocol: { type: string, example: "grpc" }
        cert_pem: { type: string, description: "گواهی نود برای pin در پنل مشتری" }
        inbounds: { type: object, description: "بخش inbounds نود برای کپی در پنل مشتری" }
        sub_token: { type: string, description: "توکن صفحه‌ی عمومی اشتراک (/sub/{token})" }
        host_info:
          type: string
          description: >
            متن آزاد اپراتور درباره‌ی هاست/کانفیگ. محتوای آن اعتبارسنجی یا محدود
            نمی‌شود، پس مصرف‌کننده باید آن را داده‌ی غیرقابل‌اعتماد در نظر بگیرد و
            پیش از نمایش، متناسب با محیط خروجی escape کند.
        note: { type: string }
    Webhook:
      type: object
      properties:
        id: { type: string }
        url: { type: string }
        events: { type: string }
        status: { type: string }